rename create-header-ui with create-profile-ui
