+++ extension.js
@@ -1,10 +1,10 @@
 import updateProfilePicker from './lib/update-profile-picker';
-import createHeaderUi from './lib/create-header-ui';
+import createProfileUi from './lib/create-profile-ui';
 import listeners from './lib/listeners';
 import model from './lib/model';
 
 document.addEventListener('DOMContentLoaded', function() {
 	updateProfilePicker();
-	createHeaderUi(model.active());
+	createProfileUi(model.active());
 	listeners();
 });
