+++ lib/delete-profile.js
@@ -1,6 +1,6 @@
 import model from './model';
 import updateProfilePicker from './update-profile-picker';
-import createHeaderUi from './create-header-ui';
+import createProfileUi from './create-profile-ui';
 import toggleUndo from './toggle-undo';
 
 
@@ -15,7 +15,7 @@ export default () => {
 		
 	model.profiles(profiles);
 	model.active(newActive);
-	createHeaderUi(newActive);
+	createProfileUi(newActive);
 	updateProfilePicker();
 	toggleUndo();
 }
