+++ lib/import-profile.js
@@ -1,5 +1,5 @@
 import model from './model';
-import createHeaderUi from './create-header-ui';
+import createProfileUi from './create-profile-ui';
 import importUi from './import-ui';
 import updateProfilePicker from './update-profile-picker';
 
@@ -19,6 +19,6 @@ export default () => {
 	profiles.push(data);
 	model.profiles(profiles);
 	model.active(newId);
-	createHeaderUi(model.active());
+	createProfileUi(model.active());
 	updateProfilePicker();
 };
