+++ lib/listeners.js
@@ -1,6 +1,6 @@
 import updateProfilePicker from './update-profile-picker';
 import createHeaderInput from './create-header-input';
-import createHeaderUi from './create-header-ui';
+import createProfileUi from './create-profile-ui';
 import toggleProfile from './toggle-profile';
 import createImportUi from './import-ui';
 import importProfile from './import-profile';
@@ -31,7 +31,7 @@ export default () => {
 			deleteProfile();
 		} else if (event.target.hasAttribute('data-delete-undo')) {
 			saveProfile(model.backup());
-			createHeaderUi(model.active())
+			createProfileUi(model.active())
 			updateProfilePicker();
 			toggleUndo(true)
 		}
@@ -39,7 +39,7 @@ export default () => {
 
 	document.querySelector('[data-banner]').addEventListener('click', (event) => {
 		if (event.target.hasAttribute('data-add-profile')) {
-			createHeaderUi();
+			createProfileUi();
 		} else if (event.target.hasAttribute('data-import-profile')) {
 			createImportUi();
 		}
