+++ lib/toggle-profile.js
@@ -1,10 +1,10 @@
 import model from './model';
-import createHeaderUi from './create-header-ui';
+import createProfileUi from './create-profile-ui';
 import updateProfilePicker from './update-profile-picker';
 
 
 export default (id) => {
-	createHeaderUi(id);
+	createProfileUi(id);
 	model.active(id);
 	updateProfilePicker();
 }
