+++ popup.html
@@ -27,8 +27,8 @@
 			<div class="profile__field">
 				Region
 				<br>
-				<label><input type="radio"> EU West (Ireland)</label>
-				<label><input type="radio"> US East (N. Virginia)</label>
+				<label><input type="radio" name="region"> EU West (Ireland)</label>
+				<label><input type="radio" name="region"> US East (N. Virginia)</label>
 			</div>
 			<label class="profile__field">Access Key ID <input></label>
 			<label class="profile__field">Secret Access Key <input></label>
